requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
Pillow